+++ src/Bore/CLI.hs
@@ -14,19 +14,30 @@ import System.Directory (canonicalizePath)
 import System.FSNotify
 import Control.Concurrent (threadDelay)
 import Control.Monad (forever)
+import Options.Applicative
 
-defaultEntryPoint :: IO ()
-defaultEntryPoint = watchServe
+{- | Determine the source and output directories.
+
+Takes two Maybe FilePaths and returns the source and output directories based on defaults
+and canonicalizes them.
+
+-}
+determineDirectories :: Maybe FilePath -> Maybe FilePath -> IO (AbsolutePath, AbsolutePath)
+determineDirectories maybeSourcePath maybeOutputPath = do
+  sourcePath <- case maybeSourcePath of
+    Just path -> canonicalizePath path
+    Nothing -> canonicalizePath defaultSourceDirectoryName
+  outputPath <- case maybeOutputPath of
+    Just path -> canonicalizePath path
+    Nothing -> canonicalizePath defaultOutputDirectoryName
+  return (sourcePath, outputPath)
 
 {- | Serve the gopherhole, if there are any changes to the children of the "source path"
 then rebuild the gopherhole.
 
 -}
-watchServe :: IO ()
-watchServe = do
-  absoluteSourcePath <- canonicalizePath defaultSourceDirectoryName
-  absoluteOutputPath <- canonicalizePath defaultOutputDirectoryName
-
+watchServe :: AbsolutePath -> AbsolutePath -> IO ()
+watchServe absoluteSourcePath absoluteOutputPath = do
   putStrLn "Building first..."
   buildTree absoluteSourcePath absoluteOutputPath
 
@@ -46,3 +57,53 @@ watchServe = do
     library <- loadOnce absoluteSourcePath
     runServerWithConfig library.config.server
     forever $ threadDelay 1000000
+
+data Command = WatchServe (Maybe FilePath) (Maybe FilePath) | Build (Maybe FilePath) (Maybe FilePath)
+
+commandParser :: Parser Command
+commandParser = subparser
+  ( command "watchServe" (info (watchServeParser <**> helper)
+      ( fullDesc
+     <> progDesc "Watch the source directory and serve the output directory" ))
+ <> command "build" (info (buildParser <**> helper)
+      ( fullDesc
+     <> progDesc "Build the output directory from the source directory" ))
+  )
+
+watchServeParser :: Parser Command
+watchServeParser = WatchServe
+  <$> optional (strOption
+      ( long "source"
+     <> metavar "SOURCE_DIR"
+     <> help "Source directory" ))
+  <*> optional (strOption
+      ( long "output"
+     <> metavar "OUTPUT_DIR"
+     <> help "Output directory" ))
+
+buildParser :: Parser Command
+buildParser = Build
+  <$> optional (strOption
+      ( long "source"
+     <> metavar "SOURCE_DIR"
+     <> help "Source directory" ))
+  <*> optional (strOption
+      ( long "output"
+     <> metavar "OUTPUT_DIR"
+     <> help "Output directory" ))
+
+defaultEntryPoint :: IO ()
+defaultEntryPoint = do
+  command' <- execParser opts
+  case command' of
+    WatchServe maybeSourcePath maybeOutputPath -> do
+      (absoluteSourcePath, absoluteOutputPath) <- determineDirectories maybeSourcePath maybeOutputPath
+      watchServe absoluteSourcePath absoluteOutputPath
+    Build maybeSourcePath maybeOutputPath -> do
+      (absoluteSourcePath, absoluteOutputPath) <- determineDirectories maybeSourcePath maybeOutputPath
+      buildTree absoluteSourcePath absoluteOutputPath
+  where
+    opts = info (commandParser <**> helper)
+      ( fullDesc
+     <> progDesc "Run the Bore CLI"
+     <> header "bore - a CLI for serving and building gopherholes" )