+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+* CLI options to just build or watch and serve. You can also override the source and
+  output directories.
+
 ## [0.1.0.0] - 2024-10-06
 
 ### Added
-- Initial release
 
-## [Unreleased]
+* Initial release
 
 [unreleased]: https://github.com/someodd/bore/compare/v0.1.0.0...HEAD
 [0.1.0.0]: https://github.com/someodd/bore/releases/tag/v0.1.0.0